requirements.txt
django==5.1
argon2-cffi==23.1.0
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
stripe==10.9.0
boto3==1.35.10
django-storages==1.14.4
django-taggit==6.0.0
django-allauth==64.2.0
django-rest-auth==0.9.5
django-crispy-forms==2.3
crispy-tailwind==1.0.3
django-ckeditor==6.7.1
django-tailwind==3.8.0
markdown==3.7
gunicorn==23.0.0
whitenoise==6.7.0
channels==4.1.0
django-redis==5.4.0
sqlparse==0.5.1
asgiref==3.8.1
pytz==2024.1
mutagen==1.47.0
django-celery-beat==2.7.0
celery==5.4.0
python-dateutil==2.9.0